:art: Rename project files
