+++ .travis.yml
@@ -5,4 +5,4 @@ before_install:
 before_script:
     - set -o pipefail
 script:
-    - xcodebuild test -project ./PxSwift.xcodeproj -scheme PxSwiftTests -configuration Debug -sdk iphonesimulator -destination 'platform=iOS Simulator,id=ABBD7191-486B-462F-80B4-AE08C5820DA1' | xcpretty -c
+    - xcodebuild test -project ./ProcessingKit.xcodeproj -scheme ProcessingKitTests -configuration Debug -sdk iphonesimulator -destination 'platform=iOS Simulator,id=ABBD7191-486B-462F-80B4-AE08C5820DA1' | xcpretty -c
