+++ ProcessingKit/ProcessingKit.h
@@ -1,0 +1,19 @@
+//
+//  ProcessingKit.h
+//  ProcessingKit
+//
+//  Created by AtsuyaSato on 2017/08/04.
+//  Copyright © 2017年 Atsuya Sato. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for ProcessingKit.
+FOUNDATION_EXPORT double ProcessingKitVersionNumber;
+
+//! Project version string for ProcessingKit.
+FOUNDATION_EXPORT const unsigned char ProcessingKitVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <ProcessingKit/PublicHeader.h>
+
+
