--- PxSwift/PxSwift.h
@@ -1,19 +1,0 @@
-//
-//  PxSwift.h
-//  PxSwift
-//
-//  Created by AtsuyaSato on 2017/08/04.
-//  Copyright © 2017年 Atsuya Sato. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for PxSwift.
-FOUNDATION_EXPORT double PxSwiftVersionNumber;
-
-//! Project version string for PxSwift.
-FOUNDATION_EXPORT const unsigned char PxSwiftVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <PxSwift/PublicHeader.h>
-
-
