--- PxSwiftExample/PxSwiftExample/Base.lproj/Main.storyboard
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="13168.3" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="R9Q-S6-WUP">
-    <device id="retina4_7" orientation="portrait">
-        <adaptation id="fullscreen"/>
-    </device>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="13147.4"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <scenes>
-        <!--View Controller-->
-        <scene sceneID="i5A-EJ-EEi">
-            <objects>
-                <viewController id="R9Q-S6-WUP" customClass="ViewController" customModule="PxSwiftExample" customModuleProvider="target" sceneMemberID="viewController">
-                    <layoutGuides>
-                        <viewControllerLayoutGuide type="top" id="xt2-P0-Rmc"/>
-                        <viewControllerLayoutGuide type="bottom" id="Zdh-Lp-VJT"/>
-                    </layoutGuides>
-                    <view key="view" contentMode="scaleToFill" id="LiO-ZE-2Pd">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <imageView userInteractionEnabled="NO" contentMode="scaleToFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="TP6-nR-kyN" customClass="CustomView" customModule="PxSwiftExample" customModuleProvider="target">
-                                <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-                            </imageView>
-                        </subviews>
-                        <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
-                        <constraints>
-                            <constraint firstAttribute="trailing" secondItem="TP6-nR-kyN" secondAttribute="trailing" id="QOg-Gn-uT7"/>
-                            <constraint firstItem="TP6-nR-kyN" firstAttribute="leading" secondItem="LiO-ZE-2Pd" secondAttribute="leading" id="hfg-rh-sRo"/>
-                            <constraint firstItem="Zdh-Lp-VJT" firstAttribute="top" secondItem="TP6-nR-kyN" secondAttribute="bottom" id="hyR-yE-rX4"/>
-                            <constraint firstItem="TP6-nR-kyN" firstAttribute="top" secondItem="LiO-ZE-2Pd" secondAttribute="top" id="jN4-z0-2LM"/>
-                        </constraints>
-                    </view>
-                    <connections>
-                        <outlet property="customView" destination="TP6-nR-kyN" id="g17-HU-XGM"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="PaG-ez-Ted" userLabel="First Responder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="72.799999999999997" y="163.26836581709148"/>
-        </scene>
-    </scenes>
-</document>
