--- PxSwiftExample/PxSwiftExample/ViewController.swift
@@ -1,25 +1,0 @@
-//
-//  ViewController.swift
-//  PxSwiftExample
-//
-//  Created by AtsuyaSato on 2017/08/05.
-//  Copyright © 2017年 Atsuya Sato. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    @IBOutlet weak var customView: CustomView!
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-         customView.isUserInteractionEnabled = true
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-}
-
