--- PxSwiftTests/PxSwiftTests.swift
@@ -1,36 +1,0 @@
-//
-//  PxSwiftTests.swift
-//  PxSwiftTests
-//
-//  Created by AtsuyaSato on 2017/08/04.
-//  Copyright © 2017年 Atsuya Sato. All rights reserved.
-//
-
-import XCTest
-@testable import PxSwift
-
-class PxSwiftTests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
-    }
-    
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-    
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-    
-}
