+++ README.md
@@ -1,3 +1,3 @@
-# PxSwift
+# ProcessingKit
 Visual Programming for iOS.
-PxSwift written in Swift and you can write like [processing](https://github.com/processing/processing).
+ProcessingKit written in Swift and you can write like [processing](https://github.com/processing/processing).
